requirements.txt
anyio==3.6.2
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
et-xmlfile==1.1.0
exceptiongroup==1.0.4
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
numpy==1.23.5
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pluggy==1.0.0
pydantic==1.10.2
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
